rewrite and add observer, and contract
