+++ demo/meson.build
@@ -1,7 +1,8 @@
 if get_option('with_demo').enabled()
     demo_apps = [
-        'json', 'csv', 'ini', 'lazy', 'fossil', 'datetime',
-        'console', 'regex', 'stream', 'thread']
+        'json', 'csv', 'ini', 'contract', 'observer',
+        'lazy', 'fossil', 'datetime', 'console',
+        'regex', 'stream', 'thread']
 
     foreach iter : demo_apps
         executable('demo_' + iter, 'demo_' + iter + '.c', dependencies : [tscl_xcore_c_dep])
